backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
pydantic[email]==2.9.2
pydantic-settings==2.6.1
openai==1.59.3
python-multipart==0.0.9
icalendar==5.0.13
pytest==8.3.4
pytest-asyncio==0.24.0
motor==3.5.1
pymongo==4.8.0
beanie==1.26.0
